+++ examples/nextjs/app/admin/plugins/[[...slug]]/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { AdminPluginsShell } from '@caspian-explorer/script-caspian-store';
+
+export default function AdminPluginsRoute() {
+  return <AdminPluginsShell />;
+}